Move types to a separate directory
