+++ RyuSocks/Packets/EndpointPacket.cs
@@ -14,6 +14,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
+using RyuSocks.Types;
 using System;
 using System.Net;
 using System.Net.Sockets;
