+++ RyuSocks/Utils/SocketErrorExtension.cs
@@ -14,6 +14,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
+using RyuSocks.Types;
 using System.Net.Sockets;
 
 namespace RyuSocks.Utils
